requirements.txt
openai
python-dotenv
feedparser
beautifulsoup4
fastapi
uvicorn[standard]
streamlit
pydantic